Split MIDI crate into two files
